feat: extend serialization/deserialization extension trait to Fq
