+++ src/lib.rs
@@ -6,7 +6,7 @@ mod element;
 mod elligator;
 mod encoding;
 mod error;
-mod fr_ext;
+mod field_ext;
 mod invsqrt;
 mod on_curve;
 mod ops;
@@ -15,7 +15,7 @@ mod sign;
 pub use element::Element;
 pub use encoding::Encoding;
 pub use error::EncodingError;
-pub use fr_ext::FrExt;
+pub use field_ext::FieldExt;
 
 pub use ark_ed_on_bls12_377::{Fq, Fr};
